Move not-common interfaces out of Interfaces.ts
